+++ src/Application/Bus/CommandDispatcher.php
@@ -12,25 +12,41 @@ declare(strict_types=1);
 
 namespace CloudCreativity\Modules\Application\Bus;
 
-use CloudCreativity\Modules\Contracts\Application\Bus\CommandHandlerContainer;
+use CloudCreativity\Modules\Contracts\Application\Bus\CommandHandlerContainer as ICommandHandlerContainer;
 use CloudCreativity\Modules\Contracts\Application\Ports\Driving\CommandDispatcher as ICommandDispatcher;
 use CloudCreativity\Modules\Contracts\Toolkit\Messages\Command;
-use CloudCreativity\Modules\Contracts\Toolkit\Pipeline\PipeContainer;
+use CloudCreativity\Modules\Contracts\Toolkit\Pipeline\PipeContainer as IPipeContainer;
 use CloudCreativity\Modules\Contracts\Toolkit\Result\Result;
 use CloudCreativity\Modules\Toolkit\Pipeline\MiddlewareProcessor;
+use CloudCreativity\Modules\Toolkit\Pipeline\PipeContainer;
 use CloudCreativity\Modules\Toolkit\Pipeline\PipelineBuilder;
+use Psr\Container\ContainerInterface;
+use ReflectionClass;
 
 class CommandDispatcher implements ICommandDispatcher
 {
+    private readonly ICommandHandlerContainer $handlers;
+
+    private readonly ?IPipeContainer $middleware;
+
     /**
      * @var array<callable|string>
      */
     private array $pipes = [];
 
     public function __construct(
-        private readonly CommandHandlerContainer $handlers,
-        private readonly ?PipeContainer $middleware = null,
+        ContainerInterface|ICommandHandlerContainer $handlers,
+        ?IPipeContainer $middleware = null,
     ) {
+        $this->handlers = $handlers instanceof ContainerInterface ?
+            new CommandHandlerContainer($handlers) :
+            $handlers;
+
+        $this->middleware = $middleware === null && $handlers instanceof ContainerInterface
+            ? new PipeContainer($handlers)
+            : $middleware;
+
+        $this->autowire();
     }
 
     /**
@@ -76,5 +92,22 @@ class CommandDispatcher implements ICommandDispatcher
         assert($result instanceof Result, 'Expecting pipeline to return a result object.');
 
         return $result;
+    }
+
+    private function autowire(): void
+    {
+        $reflection = new ReflectionClass($this);
+
+        if ($this->handlers instanceof CommandHandlerContainer) {
+            foreach ($reflection->getAttributes(WithCommand::class) as $attribute) {
+                $instance = $attribute->newInstance();
+                $this->handlers->bind($instance->command, $instance->handler);
+            }
+        }
+
+        foreach ($reflection->getAttributes(Through::class) as $attribute) {
+            $instance = $attribute->newInstance();
+            $this->pipes[] = $instance->pipe;
+        }
     }
 }