+++ src/Application/Bus/Through.php
@@ -1,0 +1,26 @@
+<?php
+
+/*
+ * Copyright 2026 Cloud Creativity Limited
+ *
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
+ */
+
+declare(strict_types=1);
+
+namespace CloudCreativity\Modules\Application\Bus;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_CLASS | Attribute::IS_REPEATABLE)]
+final readonly class Through
+{
+    /**
+     * @param non-empty-string $pipe
+     */
+    public function __construct(public string $pipe)
+    {
+    }
+}