+++ tests/Integration/Application/MathCommandBus.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * Copyright 2026 Cloud Creativity Limited
+ *
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
+ */
+
+declare(strict_types=1);
+
+namespace CloudCreativity\Modules\Tests\Integration\Application;
+
+use CloudCreativity\Modules\Application\Bus\CommandDispatcher;
+use CloudCreativity\Modules\Application\Bus\Middleware\LogMessageDispatch;
+use CloudCreativity\Modules\Application\Bus\Through;
+use CloudCreativity\Modules\Application\Bus\WithCommand;
+
+#[Through(LogMessageDispatch::class)]
+#[WithCommand(AddCommand::class, AddCommandHandler::class)]
+#[WithCommand(MultiplyCommand::class, MultiplyCommandHandler::class)]
+final class MathCommandBus extends CommandDispatcher
+{
+}